requirements.txt
langchain-openai
python-dotenv
langchain-core
streamlit
langchain
langchain-community